move ErrorCollector into errors package
