+++ packages/errors/src/index.ts
@@ -1,2 +1,3 @@
 export * from './errors';
 export * from './utils';
+export * from './ErrorCollector';
