+++ packages/project/src/index.ts
@@ -4,4 +4,3 @@ export * from './utils/convertFromV2ToV1';
 export * from './utils/v1-utils';
 export * from './utils/convertFromGgitToProject';
 export * from './legacy-specifications';
-export * from './utils/ErrorCollector';
